refactor(ui): consolidate test helpers and remove telemetry stub
Move shared utilities into helpers.py and conftest.py:
- cli_get/cli_set, SAMPLE_TEMPLATE, last_json_object → helpers.py
- activate_tab, tab_exists → DeadlineApp methods
- bundle_dir, submitter_env, gui_submit fixtures → conftest.py

Remove the get_account_id stub from the sitecustomize shim. After
the remove-sts merge (78b10da), telemetry no longer calls STS inline
or prints to stdout on failure, so the stub is unnecessary. Telemetry
is already disabled via DEADLINE_CLOUD_TELEMETRY_OPT_OUT=true.

Net reduction of ~350 lines across test files.
