+++ test/ui/test_config_gui_log_level.py
@@ -2,10 +2,6 @@
 
 """UI tests for ``deadline config gui`` — log level setting.
 
-Launches the real GUI as a subprocess and drives it through xa11y. The
-CLI is pointed at an in-process MockDeadlineBackend via ``deadline_env``
-so no real AWS calls occur.
-
 Note: on macOS, Qt combo-box popup items cannot be reliably activated
 via the accessibility API, so we mutate the config value with
 ``deadline config set`` and only assert the GUI reflects it.
@@ -13,26 +9,9 @@ via the accessibility API, so we mutate the config value with
 
 from __future__ import annotations
 
-import subprocess
-
 import pytest
 
-from helpers import ConfigDialog
-
-
-def _cli_get(env: dict, setting: str) -> str:
-    return subprocess.check_output(
-        ["deadline", "config", "get", setting], env=env, text=True, stderr=subprocess.DEVNULL
-    ).strip()
-
-
-def _cli_set(env: dict, setting: str, value: str) -> None:
-    subprocess.check_call(
-        ["deadline", "config", "set", setting, value],
-        env=env,
-        stdout=subprocess.DEVNULL,
-        stderr=subprocess.DEVNULL,
-    )
+from helpers import ConfigDialog, cli_get, cli_set
 
 
 class TestConfigGuiOpens:
@@ -54,7 +33,7 @@ class TestConfigGuiOpens:
 
     def test_shows_current_log_level(self, deadline_env):
         _, env = deadline_env
-        expected = _cli_get(env, "settings.log_level")
+        expected = cli_get(env, "settings.log_level")
         with ConfigDialog.open(env=env) as app:
             assert app.log_level == expected
 
@@ -72,7 +51,7 @@ class TestLogLevelRoundTrip:
     @pytest.mark.parametrize("level", ["ERROR", "WARNING", "INFO", "DEBUG"])
     def test_each_level_displays_correctly(self, deadline_env, level: str):
         _, env = deadline_env
-        _cli_set(env, "settings.log_level", level)
+        cli_set(env, "settings.log_level", level)
         with ConfigDialog.open(env=env) as app:
             assert app.log_level == level
 
@@ -80,21 +59,21 @@ class TestLogLevelRoundTrip:
 class TestCancelDoesNotSave:
     def test_cancel_leaves_config_unchanged(self, deadline_env):
         _, env = deadline_env
-        _cli_set(env, "settings.log_level", "DEBUG")
+        cli_set(env, "settings.log_level", "DEBUG")
         with ConfigDialog.open(env=env) as app:
             assert app.log_level == "DEBUG"
             app.close("Cancel")
-        assert _cli_get(env, "settings.log_level") == "DEBUG"
+        assert cli_get(env, "settings.log_level") == "DEBUG"
 
 
 class TestOkSaves:
     def test_ok_persists_value(self, deadline_env):
         _, env = deadline_env
-        _cli_set(env, "settings.log_level", "INFO")
+        cli_set(env, "settings.log_level", "INFO")
         with ConfigDialog.open(env=env) as app:
             assert app.log_level == "INFO"
             app.close("Ok")
-        assert _cli_get(env, "settings.log_level") == "INFO"
+        assert cli_get(env, "settings.log_level") == "INFO"
 
 
 class TestReopenAfterCancel:
@@ -110,7 +89,7 @@ class TestReopenAfterCancel:
 class TestChangeOkReopen:
     def test_change_persists_across_reopens(self, deadline_env):
         _, env = deadline_env
-        _cli_set(env, "settings.log_level", "DEBUG")
+        cli_set(env, "settings.log_level", "DEBUG")
         with ConfigDialog.open(env=env) as app:
             assert app.log_level == "DEBUG"
             app.close("Ok")
